asm/f32:  Adding stubs for dot asm functions.
